+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # pygrader
 
+
 ## 1.7.0
 
 - Pylintrc can now be part of the config
+- Add total score and max score to JSON, CSV, and PlainText result displays
 
 ## 1.6.0