+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # pygrader
 
-
 ## 1.7.0
 
 - Pylintrc can now be part of the config
 - Add total score and max score to JSON, CSV, and PlainText result displays
+- Environment variables can now be passed from the config
 
 ## 1.6.0